+++ test/decorators/noDispatch.tests.ts
@@ -1,26 +1,32 @@
 import { expect } from 'chai';
 import { component, noDispatch, ReduxApp } from 'src';
 
+// tslint:disable:no-unused-expression
+
 describe(nameof(noDispatch), () => {
 
-    it("non-decorated method affects the store state", () => {
+    it("non-decorated method dispatches an action", () => {
 
         @component
         class App {
-            
+
             public value = 0;
 
             public increment() {
                 this.value = this.value + 1;
             }
         }
-        const app = new ReduxApp(new App());
-        app.root.increment();
 
-        expect(app.store.getState().value).to.eql(1);
+        var dispatched = false;
+
+        const app = new ReduxApp(new App());
+        app.store.subscribe(() => dispatched = true);
+
+        app.root.increment();
+        expect(dispatched).to.be.true;
     });
 
-    it("decorated method does not affect the store state", () => {
+    it("decorated method does not dispatch an action", () => {
 
         @component
         class App {
@@ -31,10 +37,14 @@ describe(nameof(noDispatch), () => {
                 this.value = this.value + 1;
             }
         }
-        const app = new ReduxApp(new App());
-        app.root.increment();
 
-        expect(app.store.getState().value).to.eql(0);
+        var dispatched = false;
+
+        const app = new ReduxApp(new App());
+        app.store.subscribe(() => dispatched = true);
+
+        app.root.increment();
+        expect(dispatched).to.be.false;
     });
 
 });