+++ test/testTypes.ts
@@ -1,13 +1,4 @@
-import { component } from 'src';
 import { Store, Unsubscribe } from 'redux';
-
-@component
-export class Counter {
-    public value = 0;
-    public increment() {
-        this.value = this.value + 1;
-    }
-}
 
 export class FakeStore implements Store<any> {
     public dispatch(): any {