+++ CHANGELOG.md
@@ -2,9 +2,20 @@
 
 ## [Unreleased](https://github.com/alonrbar/redux-app/tree/develop)
 
+### Added
+
+- New decorator: computed.
+
 ### Changed
 
 - When updating state do not remove component properties that exists on the state but are undefined.
+- The 'updateState' option is no longer a component level option but only a global option.
+
+### Fixed
+
+- Major refactor to the library's internals that address several issues. The primary issues where:
+  1. All patched methods where store on Component prototype. This would mean that two different classes having a method with the same name will dispatch the same action.
+  2. Updates to components nested inside regular objects where not applied correctly.
 
 ## [1.4.0 - 2017-10-22](https://github.com/alonrbar/redux-app/tree/v1.4.0)